Rename SkinChanger.h to InventoryChanger.h
