+++ Osiris/Hacks/InventoryChanger.h
